[Fw] Splitted source files into MEMS and EEPROM folders
