+++ MEMS/ADXL345.c
